Move design logic to dedicated +Design file
