+++ RoomScan/Sources/RoomScan/Landing/Components/RoomScanCell/RoomScanCell+Design.swift
@@ -1,0 +1,25 @@
+import UIKit
+import SnapKit
+
+extension RoomScanCell {
+
+    public func buildViews() {
+        createViews()
+        styleViews()
+        defineLayoutForViews()
+    }
+
+    public func createViews() {
+        cardView = RoomScanCardView()
+        addSubview(cardView)
+    }
+
+    public func styleViews() {}
+
+    public func defineLayoutForViews() {
+        cardView.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
+    }
+
+}
