+++ RoomScan/Sources/RoomScan/Landing/Components/RoomScanCell/RoomScanCell.swift
@@ -37,26 +37,3 @@ class RoomScanCell: UICollectionViewCell {
     }
 
 }
-
-extension RoomScanCell {
-
-    public func buildViews() {
-        createViews()
-        styleViews()
-        defineLayoutForViews()
-    }
-
-    public func createViews() {
-        cardView = RoomScanCardView()
-        addSubview(cardView)
-    }
-
-    public func styleViews() {}
-
-    public func defineLayoutForViews() {
-        cardView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-    }
-
-}
