+++ VirtualObject/Sources/VirtualObject/Landing/Components/VirtualObjectCell/CardView/VirtualObjectCardView.swift
@@ -1,6 +1,6 @@
 import UIKit
 import SceneKit
-import SnapKit
+import Core
 
 class VirtualObjectCardView: UIView {
 
@@ -23,50 +23,12 @@ class VirtualObjectCardView: UIView {
 
     func set(type: VirtualObjectType) {
         titleLabel.text = type.title
-        guard let url = Bundle.module.url(forResource: type.rawValue, withExtension: "usdz") else { return }
+
+        guard
+            let url = Bundle.module.url(forResource: type.rawValue, withExtension: FileType.usdz.fileExtension)
+        else { return }
 
         imageView.load3DModelThumbnail(from: url)
-    }
-
-}
-
-extension VirtualObjectCardView {
-
-    func buildViews() {
-        createViews()
-        styleViews()
-        defineLayoutForViews()
-    }
-
-    func createViews() {
-        stackView = UIStackView()
-        addSubview(stackView)
-
-        imageView = UIImageView()
-        stackView.addArrangedSubview(imageView)
-
-        titleLabel = UILabel()
-        stackView.addArrangedSubview(titleLabel)
-    }
-
-    func styleViews() {
-        backgroundColor = .white
-        roundAllCorners(withRadius: cornerRadius)
-        addBorder(width: 2, color: .black)
-
-        stackView.axis = .vertical
-        stackView.alignment = .center
-
-        imageView.contentMode = .scaleAspectFit
-
-        titleLabel.textColor = .black
-        titleLabel.font = UIFont(with: .futura, size: 18)
-    }
-
-    func defineLayoutForViews() {
-        stackView.snp.makeConstraints {
-            $0.edges.equalToSuperview().inset(defaultPadding)
-        }
     }
 
 }
