+++ VirtualObject/Sources/VirtualObject/Landing/Components/VirtualObjectCell/VirtualObjectCell+Design.swift
@@ -1,0 +1,24 @@
+import SnapKit
+
+extension VirtualObjectCell {
+
+    public func buildViews() {
+        createViews()
+        styleViews()
+        defineLayoutForViews()
+    }
+
+    public func createViews() {
+        cardView = VirtualObjectCardView()
+        addSubview(cardView)
+    }
+
+    public func styleViews() {}
+
+    public func defineLayoutForViews() {
+        cardView.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
+    }
+
+}
